Finish
Decided to leave it here for the moment.

Still need to implement localStorage but it seems impossible to do so
with methods.

The problem in this case is that
localStorage.setItem("list", JSON.stringify(list)) saves the methods of
List but it doesn't save the methods of either Task or Project and that
breaks the code.